+++ packages/palette/src/elements/data-visualization/BarChart/BarChart.test.tsx
@@ -3,8 +3,8 @@ import "jest-styled-components"
 import React from "react"
 import { color } from "../../../helpers"
 import { Theme } from "../../../Theme"
+import { Sans } from "../../../tokens"
 import { Flex } from "../../layout"
-import { Sans } from "../../Typography"
 import { BarChart, BarChartProps } from "../BarChart"
 import { Bar } from "./Bar"
