+++ packages/palette/src/elements/data-visualization/BarChart/BarChart.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useRef, useState } from "react"
 import styled from "styled-components"
 import { color } from "../../../helpers"
+import { Sans } from "../../../tokens"
 import { Flex } from "../../layout"
-import { Sans } from "../../Typography"
 import { Bar } from "./Bar"
 import { BarLabel, BarLabelProps, isBarLabelProps } from "./BarLabel"
 import { ProvideMousePosition } from "./MousePositionContext"
