+++ packages/palette/src/elements/dialogs/Message/Message.tsx
@@ -2,8 +2,8 @@ import React, { SFC } from "react"
 import { color } from "../../../helpers"
 import { styledWrapper } from "../../../platform/primitives"
 import { SansSize } from "../../../Theme"
+import { Sans } from "../../../tokens"
 import { Flex, FlexProps } from "../../layout"
-import { Sans } from "../../Typography"
 
 /**
  * Spec: zpl.io/2Zg4Rdq
