+++ packages/palette/src/tokens/index.tsx
@@ -1,2 +1,3 @@
 export * from "./Colors"
 export * from "./Grid"
+export * from "./Typography"
